requirement.txt
asgiref==3.8.1
Django==4.2.17
djangorestframework==3.15.2
pillow==11.1.0
sqlparse==0.5.3
typing-extensions==4.12.2
tzdata==2024.2
whitenoise==6.6.0 #must be in requirements.txt
gunicorn==21.2.0 #must be in requirements.txt